// lib/types.ts
export type HospitalTypeKey =
  | 'nursing'       // 요양병원
  | 'psychiatric'   // 정신병원
  | 'rehabilitation'// 재활의료기관
  | 'acute'         // 급성기병원
  | 'tertiary'      // 상급종합병원
  | 'general'       // 종합병원
  | 'hospital'      // 병원
  | 'dental'        // 치과병원
  | 'korean'        // 한방병원
  | 'other'         // 기타병원
  | 'custom';       // 직접입력